Refactor OAuth flows into separate controllers
