+++ app/views/oauth/client_error.html.erb
@@ -5,7 +5,7 @@
       <%= tag.p t('.lede'), class: 'mb-2' %>
       <div class="text-center text-red-800 bg-red-200 rounded border-1 border-red-800 p-4">
         <%= tag.h3 error_class, class: 'text-xl semibold' %>
-        <%= tag.p message %>
+        <%= tag.p error_message %>
       </div>
     </div>
   </div>
